requirements.txt
# Core Dependencies
gradio>=4.0.0
transformers>=4.35.0
torch>=2.0.0
torchaudio>=2.0.0

# Speech Recognition
openai-whisper>=20231117
soundfile>=0.12.1

# NLP and Text Processing
sentencepiece>=0.1.99
protobuf>=3.20.0
accelerate>=0.24.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
python-dotenv>=1.0.0

# API Clients
groq>=0.32.0

# PDF Generation
reportlab>=4.4.4
